ai_review/requirements.txt
aio-pika>=9.4.0
google-cloud-storage==2.18.2
google-genai>=1.0.0
python-dotenv==1.0.1
pypdf==5.1.0
python-docx==1.1.2
pillow==10.4.0
opencv-python-headless==4.10.0.84
